remove highlighting of copyright lines in comments
Only the first year in a list of years is currently highlighted.  Fixing
this to match all years is easy, however, it should probably extend to
the end of the copyright line which is impossible to reliably match.

Matching copyright lines is sufficiently unusual for a distributed
syntax file that it seems reasonable to just remove it.